Simplify decoding of data in file handle
There was some code in _go_to_gi which tried to set a mode based on
the itemtype, but then the mode wasn't used and "rb" was used for
send_selector, which makes sense because we now handle multiple
character encodings and thus we always expect a binary file, for the
exception of queries. Thus, we now call _decode_text for itemtype "7"
and for do_veronica.