Catch OSError, too
On this other laptop, connecting to a site with no TSL support but
VF-1 in battloid mode gives me a new error... The list of errors is
amazingly wide. Not sure if OSError includes any interesting
information. The backtrace shows that the error happens during
handshake, so I just mention the tls command in the reply. Better than
crashing.