Destroy beauty in name of practicality.
Dump formatted (i.e. human friendly) menus to a separate tempfile
to the tempfile used for text content.  This lets type 1 items be
run through 'less' correctly, but also allows the "source" for the
menu to be saved as-is via 'save'.

Note that the (thus far undocumented) auto_page option has changed
its functionality a bit because of this.